% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCalendarLikeMicrosoft.R
\name{createCalendarLikeMicrosoft}
\alias{createCalendarLikeMicrosoft}
\title{Create a calendar with weeks that match what the WEEKNUM([Date]) and DATEPART(ww, [Date]) functions return in Excel and sQL, respectively.}
\usage{
createCalendarLikeMicrosoft(startYear, yearSubGroup)
}
\arguments{
\item{startYear}{is a numeric value that is a real year}

\item{yearSubGroup}{is a sub-group of a year, i.e. 'Week', 'Month', or 'Quarter'}
}
\value{
a data frame of dates starting at Jan 1 of the startYear and ending today, which the corresponding year and DateGroup (e.g. 2015-01 for Week 1 of 2015 if yearSubGroup = 'Week')
}
\description{
This function takes a starting year and a sub-group of year (i.e. week, month, quarter) and returns every date between that start year and today with a DateGroup.
}
\details{
This function takes in a startYear and yearSubGroup and returns a data frame containing all the dates from the start of the startYear
through today along with the year and yearSubGroup number. If the yearSubGroup is 'Week', then the function will return week numbers consistent
with Microsoft's Excel WEEKNUM and SQL DATEPART(ww, ) functions.
}
\author{
Aimie Faucett
}
