% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEvenWeeks.R
\name{createEvenWeek}
\alias{createEvenWeek}
\title{Create weeks that are relatively even such that there are 52 weeks in the year}
\usage{
createEvenWeek(calFrame)
}
\arguments{
\item{calFrame}{a data frame containing a column of consecutive dates}
}
\value{
a data frame that contains the date and the corresponding even week
}
\description{
This function takes in a calendar data frame that has an entry for every date since a starting point up through today and
returns a data frame that has a week number 1-52 that are approximately even starting with Jan 1 as Week 1, day 1.
}
\details{
This function takes a calFrame and returns a data.frame that contains the date with its corresponding CDC reporting week
}
\author{
Aimie Faucett
}
