% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeeksLikeCDC.R
\name{createWeeksLikeCDC}
\alias{createWeeksLikeCDC}
\title{Create weeks similar to CDC reporting periods}
\usage{
createWeeksLikeCDC(calFrame)
}
\arguments{
\item{calFrame}{a data frame containing a column of consecutive dates}
}
\value{
a data frame that contains the date and the corresponding CDC reporting week
}
\description{
This function takes in a calendar data frame that has an entry for every date since a starting point up through today and
returns a data frame that has a week number matching CDC reporting for each assoicated date.
}
\details{
This function takes a calFrame and returns a data.frame that contains the date with its corresponding CDC reporting week
}
\author{
Aimie Faucett
}

