% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStartDate.R
\name{findStartDate}
\alias{findStartDate}
\title{Find the start date in the format of a character DateGroup using today's DateGroup as the starting point.}
\usage{
findStartDate(msCalFrame, yearSubGroup, timeUnitsToKeep, rollPeriods = 0)
}
\arguments{
\item{msCalFrame}{is a data frame (non-sparse) of all dates and DateGroups, an output of \code{createCalendarLikeMicrosoft} function}

\item{yearSubGroup}{is a sub-group of a year, i.e. 'Week', 'Month', or 'Quarter'}

\item{timeUnitsToKeep}{is a numeric value indicating how many of DateGroup entries to keep (incremented by 1 backwards from today's DateGroup)}

\item{rollPeriods}{is a numeric value indicating how many additional periods should be included in case the user wants to find a rolling sum, the default is 0.}
}
\value{
a character string DateGroup at timeUnitsToKeep + rollPeriods prior (e.g. 2016-01 if today's DateGroup is 2016-10 and timeUnitsToKeep = 9 and rollPeriods = 0)
}
\description{
The function sequences DateGroups from last date to today, incrementing by -1, and keeps the number of DateGroup units as specified by the user.
}
\details{
This function takes in msCalFrame, yearSubGroup, timeUnitsToKeep, and rollPeriods. The unique msCalFrame DateGroups are found and then sequenced in descending order. The number of entries
kept is specified by timeUnitsToKeep + rollPeriods. The function returns the startDate (character string) as determined by the timeUnitsToKeep +
rollPeriods prior to the most current DateGroup.
}
\author{
Aimie Faucett
}

