% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCalSparseFrames.R
\name{mergeCalSparceFrames}
\alias{mergeCalSparceFrames}
\title{Merge data frames with date-sparse data}
\usage{
mergeCalSparceFrames(numFrame, denomFrame, mergeColsX, mergeColsY, numNumeric,
  denomNumeric, sparseHandling, rollPeriods = 0)
}
\arguments{
\item{numFrame}{a sparse data frame created using the \code{aggregateAndFillDateGroupGaps} function}

\item{denomFrame}{a sparse data frame created using the \code{aggregateAndFillDateGroupGaps} function}

\item{mergeColsX}{a vector of character strings specifying which columns to use in the \code{\link[base]{merge}} function's by.x parameter}

\item{mergeColsY}{a vector of character strings specifying which columns to use in the \code{\link[base]{merge}} function's by.y parameter}

\item{numNumeric}{a character string that identifies which column of the numFrame contains the numeric value to use as the numerator in the rate calculation}

\item{denomNumeric}{a character string that identifies which column of the denomFrame contains the numeric value to use as the denominator in the rate calculation}

\item{sparseHandling}{a numeric (either NA or 0) that tells the function how to handle sparse Rate values after the merge}

\item{rollPeriods}{a numeric (default = 0) that indicates the number of periods to include in the rolling rate (0, default, is not rolled)}
}
\value{
a data frame that contains the orginal columns in numFrame as well as a Rate that is calculated using the numNumeric/denomNumeric and adjusted using sparseHandling for DateGroups that are sparse in the numFrame
}
\description{
This function merges two data frames that are created using aggregateAndFillDataGroupGaps and creates a rate with roll capability and sparse entry handling.
}
\details{
This function takes a numFrame, denomFrame, mergeColsX, mergeColsY, numNumeric, denomNumeric, sparseHandling, rollPeriods, and lagPeriods. It
returns a data.frame that contains a DateGroup, columns from numFrame (less the numNumeric column), and a Rate (either rolled or not with
sparse entries handled as specified by the sparseHandling parameter). The rate is calculated by merging the numFrame and denomFrame by the
mergeColsX and mergeColsY and then finding a rate for each unique combination of the merged columns.
}
\author{
Aimie Faucett
}

