% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformToEpiWeeks.R
\name{transformToEpiWeeks}
\alias{transformToEpiWeeks}
\title{Create EPI weeks}
\usage{
transformToEpiWeeks(calFrame)
}
\arguments{
\item{calFrame}{a data frame containing a column of consecutive dates produced using createCalendarLikeMicrosoft with 'Week' as the yearSubGroup}
}
\value{
a data frame that contains the date, EPI year and EPI week
}
\description{
This function takes in a calendar data frame that has an entry for every date since a starting point up through today and
returns a data frame that has an EpiWeek and EpiYear along with the Date.
}
\details{
This function takes a calFrame and returns a data.frame that contains the date with its corresponding EPI reporting week
}
\author{
Aimie Faucett
}

