% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStatsToSparseHandledData.R
\name{addStatsToSparseHandledData}
\alias{addStatsToSparseHandledData}
\title{Return the mean and limits for the sparse data frame, allowing for alternate limits and one sided or two sided limits.}
\usage{
addStatsToSparseHandledData(sparseRateFrame, ratePartitionVec,
  returnLimits = FALSE, limitFactor = 3, limitSide = "upper",
  altUL = 0.001, altLL = 0)
}
\arguments{
\item{sparseRateFrame}{a sparse data frame created using the \code{mergeCalSparseFrames} function}

\item{ratePartitionVec}{a character string vector that specifies how to parition the categories in the data frame}

\item{returnLimits}{a logical with default FALSE that specifies whether or not the data frame returned should have limits}

\item{limitFactor}{a numeric, default of 3, that specifies the number of standard deviations from the mean to set the limit}

\item{limitSide}{a character string, default of upper, that indicates whether the limit should be +X from the mean (upper), -X from the mean (lower), or +/-X from the mean (two.sided)}

\item{altUL}{a numeric, default of 0.001, that indicates a default limit if the altUL > limitFactor*sd(X)}

\item{altLL}{a numeric, default of 0, that indicates a default limit if the altLL < limitFactor*sd(X)}
}
\value{
a data frame that contains the orginal columns in sparseRateFrame, as well as an average for each of the ratePartitionVec combinations. It may also have limits if limit parameters are specified.
}
\description{
This function takes a data frame outputted from the mergeCalSparseFrames function and paritions the data frame by the rateParitionVec parameter. It then finds the mean and limits, and colors data points accordingly.
}
\details{
This function takes in a sparseRateFrame that is made using the \code{mergeCalSparseFrames} function and return the mean for each unique combination of
the entries in the ratePartitionVec. If the limit parameters are entered, the Rate variable is colored as 'pass' or 'review' based on limitFactor*sd(X)
and that value is either added, subtracted, or added and subtracted to make a range about the mean where the data set falls in or out of teh limits. The
altUL and altLL allow the user to specify an alternate limit that can be applied instead of the limitFactor.
}
\author{
Aimie Faucett
}

